flattenObject fix
flattenObject is flattening empty arrays in the wrong way, the result of
it is that empty arrays appear to be simple objects. The reason this is
happening is because we're trusting on `dotJoin`, which uses `compact`,
which by definition removes zeroes from the array.

My solution is simply to add another `dotJoin` that allows zeroes. I'm
naming it `dotJoinWith`, since it accepts a function that will be used
to select how to filter the array.